refactor drechtsteden-models: split into category‑specific sealed subclasses
- extract RES region, municipalities, business parks, and neighborhoods into sealed classes extending 'DrechtstedenTwinModel'

-keep shared URL logic in base; no functional changes
