+++ site/src/jsMain/kotlin/com/zenmo/web/zenmo/domains/lux/subdomains/drechtsteden/pages/municipalities/HardinxveldPage.kt
@@ -1,29 +1,16 @@
 package com.zenmo.web.zenmo.domains.lux.subdomains.drechtsteden.pages.municipalities
 
 import androidx.compose.runtime.Composable
-import com.varabyte.kobweb.compose.foundation.layout.Column
-import com.varabyte.kobweb.compose.ui.Alignment
-import com.varabyte.kobweb.compose.ui.Modifier
-import com.varabyte.kobweb.compose.ui.modifiers.fillMaxWidth
-import com.varabyte.kobweb.compose.ui.modifiers.gap
-import com.varabyte.kobweb.compose.ui.modifiers.margin
-import com.varabyte.kobweb.silk.components.layout.SimpleGrid
-import com.varabyte.kobweb.silk.components.layout.numColumns
-import com.zenmo.web.zenmo.components.widgets.ZenmoInlineLink
 import com.zenmo.web.zenmo.domains.lux.components.ProfileContactCard
-import com.zenmo.web.zenmo.domains.lux.components.VerticalProfileContactCard
-import com.zenmo.web.zenmo.domains.lux.components.model.DrechtstedenTwinModel
+import com.zenmo.web.zenmo.domains.lux.components.model.DrechtstedenMunicipality
 import com.zenmo.web.zenmo.domains.lux.components.model.ModelPageContent
 import com.zenmo.web.zenmo.domains.lux.subdomains.drechtsteden.components.ModelInDevelopmentInfoWidget
 import com.zenmo.web.zenmo.domains.lux.subdomains.drechtsteden.components.layout.DrechtstedenTwinLayout
-import com.zenmo.web.zenmo.domains.lux.widgets.headings.HeaderText
-import com.zenmo.web.zenmo.domains.lux.widgets.headings.SubHeaderText
 import com.zenmo.web.zenmo.domains.zenmo.sections.team.ZenmoTeam
-import org.jetbrains.compose.web.css.cssRem
 
 @Composable
 fun HardinxveldPage() {
-    val twinModel = DrechtstedenTwinModel.SliedrechtMunicipality
+    val twinModel = DrechtstedenMunicipality.SliedrechtMunicipality
     DrechtstedenTwinLayout(
         title = twinModel.title,
         useTwinPageHeader = true,
@@ -55,44 +42,6 @@ fun HardinxveldPage() {
                 telephoneNumber = "+31 6 14910380",
                 enSubtitle = "Website and model development",
                 nlSubtitle = "Website en model ontwikkeling",
-            )
-        }
-    }
-}
-
-
-// should we have a ContactCardPair component? instead just one for both Auke and Naud?
-@Composable
-fun AukeNaudtContactCard() {
-    Column(
-        modifier = Modifier.fillMaxWidth().gap(1.cssRem),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        HeaderText(
-            enText = "Contact and info", nlText = "Contact en info", modifier = Modifier.margin(0.cssRem)
-        )
-        SubHeaderText(
-            enText = "Website and model development",
-            nlText = "Website en model ontwikkeling",
-            modifier = Modifier.margin(0.cssRem)
-        )
-        ZenmoInlineLink()
-        SimpleGrid(
-            numColumns = numColumns(base = 1, md = 2, lg = 2, xl = 2),
-            modifier = Modifier.gap(4.cssRem).margin(topBottom = 3.cssRem)
-        ) {
-            VerticalProfileContactCard(
-                name = ZenmoTeam.NAUD_LOOMANS.memberName,
-                imageUrl = ZenmoTeam.NAUD_LOOMANS.image,
-                email = ZenmoTeam.NAUD_LOOMANS.email,
-                telephoneNumber = "+31 (6) 1524 9535",
-            )
-
-            VerticalProfileContactCard(
-                name = ZenmoTeam.AUKE.memberName,
-                imageUrl = ZenmoTeam.AUKE.image,
-                email = ZenmoTeam.AUKE.email,
-                telephoneNumber = "+31 (6) 5161 4294",
             )
         }
     }
