+++ site/src/jsMain/kotlin/com/zenmo/web/zenmo/domains/lux/subdomains/drechtsteden/pages/resneighborhoods/OverTSpoor.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.Composable
 import com.zenmo.web.zenmo.components.widgets.SectionContainer
 import com.zenmo.web.zenmo.domains.lux.components.ModelWrapper
 import com.zenmo.web.zenmo.domains.lux.components.ProfileContactCard
-import com.zenmo.web.zenmo.domains.lux.components.model.DrechtstedenTwinModel
+import com.zenmo.web.zenmo.domains.lux.components.model.DrechtstedenResNeighborhood
 import com.zenmo.web.zenmo.domains.lux.sections.LuxSectionContainerStyleVariant
 import com.zenmo.web.zenmo.domains.lux.subdomains.drechtsteden.components.layout.DrechtstedenTwinLayout
 import com.zenmo.web.zenmo.domains.zenmo.sections.team.ZenmoTeam
@@ -12,7 +12,7 @@ import com.zenmo.web.zenmo.domains.zenmo.sections.team.ZenmoTeam
 
 @Composable
 fun OverTSpoor() {
-    val twinModel = DrechtstedenTwinModel.OverTSpoor
+    val twinModel = DrechtstedenResNeighborhood.OverTSpoor
     DrechtstedenTwinLayout(
         title = twinModel.title,
         useTwinPageHeader = true,
