+++ site/src/jsMain/kotlin/com/zenmo/web/zenmo/domains/lux/subdomains/drechtsteden/pages/resneighborhoods/ResNeighborhoodsRouting.kt
@@ -1,7 +1,7 @@
 package com.zenmo.web.zenmo.domains.lux.subdomains.drechtsteden.pages.resneighborhoods
 
 import com.varabyte.kobweb.core.init.InitKobwebContext
-import com.zenmo.web.zenmo.domains.lux.components.model.DrechtstedenTwinModel
+import com.zenmo.web.zenmo.domains.lux.components.model.DrechtstedenResNeighborhood
 
 
 fun resNeighborhoodsRouting(
@@ -9,7 +9,7 @@ fun resNeighborhoodsRouting(
 ) {
     val path = "/res-neighborhoods"
     with(ctx.router) {
-        register("$path/${DrechtstedenTwinModel.Oostdonk.name}") { Oostdonk() }
-        register("$path/${DrechtstedenTwinModel.OverTSpoor.name}") { OverTSpoor() }
+        register("$path/${DrechtstedenResNeighborhood.Oostdonk.name}") { Oostdonk() }
+        register("$path/${DrechtstedenResNeighborhood.OverTSpoor.name}") { OverTSpoor() }
     }
 }
